pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Julien Gaye <julien.gaye@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.1"
neo4j = "^5.12.0"
uvicorn = "^0.23.2"
pyyaml = "^6.0.1"
pydantic-settings = "^2.3.4"
types-pyyaml = "^6.0.12.20240311"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
httpx = "^0.27.0"
pytest-cov = "^5.0.0"
black = "^24.4.2"
pytest-dotenv = "^0.5.2"
pre-commit = "^3.7.1"
mypy = "^1.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
